+++ src/Repository.ts
@@ -3,13 +3,14 @@ import type { ParamMap } from './types'
 export interface Repository<Type> {
 	read(
 		params: ParamMap,
-		options?: { key?: string | number | boolean; } & Record<string, unknown>,
+		options?: { key?: string | number | boolean } & Record<string, unknown>,
 	): {
 		responsePromise: Promise<{
 			ok: boolean
 			aborted?: boolean
 			abortReason?: string
 			data?: Type[]
+			item?: Type
 			metadata?: ParamMap
 		}>
 		abort?: (reason?: string) => void
@@ -17,23 +18,7 @@ export interface Repository<Type> {
 	}
 
 	create(
-		item: Type,
-		params?: ParamMap,
-		options?:  Record<string, unknown>,
-	): {
-		responsePromise: Promise<{
-			ok: boolean
-			aborted?: boolean
-			abortReason?: string
-			data?: Type
-			metadata?: ParamMap
-		}>
-		abort?: (reason?: string) => void
-		signal?: AbortSignal
-	}
-
-	update(
-		item: Type,
+		payload: Type | Type[] | undefined,
 		params?: ParamMap,
 		options?: Record<string, unknown>,
 	): {
@@ -41,7 +26,25 @@ export interface Repository<Type> {
 			ok: boolean
 			aborted?: boolean
 			abortReason?: string
-			data?: Type
+			data?: Type[]
+			item?: Type
+			metadata?: ParamMap
+		}>
+		abort?: (reason?: string) => void
+		signal?: AbortSignal
+	}
+
+	update(
+		payload: Type | Type[] | undefined,
+		params?: ParamMap,
+		options?: Record<string, unknown>,
+	): {
+		responsePromise: Promise<{
+			ok: boolean
+			aborted?: boolean
+			abortReason?: string
+			data?: Type[]
+			item?: Type
 			metadata?: ParamMap
 		}>
 		abort?: (reason?: string) => void
@@ -49,6 +52,7 @@ export interface Repository<Type> {
 	}
 
 	remove(
+		payload: unknown,
 		params: ParamMap,
 		options?: Record<string, unknown>,
 	): {